Added the ability to show the disposition on the contact event view.
 * refactored the show view for contact event

 * broke out 'show'ing the demographics tab into partials

 * updated UAT's to reflect current behavior of the UI.

 * added default codes back into codes.yml for new installs

 * updated the migration to only add default codes if in production

 * tri446

 * csv export stuff to follow

git-svn-id: https://trisano.csinitiative.net/svn/trisano/top-level/trunk@1309 b21a84b9-e453-0410-92a3-c4f515f7ae45

+++ trisano/webapp/db/migrate/064_add_contact_disposition_to_people.rb
@@ -23,19 +23,25 @@ class AddContactDispositionToPeople < ActiveRecord::Migration
   def self.up
     add_column(:people, :disposition_id, :integer)    
     
-    [{:code_description => 'Preventative treatment', :the_code => 'PT'},
-     {:code_description => 'Refused preventative treatment', :the_code => 'RPT'},
-     {:code_description => 'Infected, brought to treatment', :the_code => 'IBT'},
-     {:code_description => 'Infected, not treated', :the_code => 'INT'},
-     {:code_description => 'Previously treated for this infection', :the_code => 'PTFTI'},
-     {:code_description => 'Not infected', :the_code => 'NI'},
-     {:code_description => 'Insufficient information to begin investigation', :the_code => 'IIBI'},
-     {:code_description => 'Unable to locate', :the_code => 'UTL'},
-     {:code_description => 'Located, refused exam and/or treament', :the_code => 'LR'},
-     {:code_description => 'Out of jurisdiction', :the_code => 'OOJ'},
-     {:code_description => 'Other', :the_code => 'O'}
-    ].each_with_index do |type, i|
-      ExternalCode.create!(type.merge(:sort_order => i*10, :code_name => 'contactdispositiontype'))
+    if RAILS_ENV == 'production'
+      transaction do
+        say 'Adding defaault contact dispostion codes to external_codes table'
+
+        [{:code_description => 'Preventative treatment', :the_code => 'PT'},
+         {:code_description => 'Refused preventative treatment', :the_code => 'RPT'},
+         {:code_description => 'Infected, brought to treatment', :the_code => 'IBT'},
+         {:code_description => 'Infected, not treated', :the_code => 'INT'},
+         {:code_description => 'Previously treated for this infection', :the_code => 'PTFTI'},
+         {:code_description => 'Not infected', :the_code => 'NI'},
+         {:code_description => 'Insufficient information to begin investigation', :the_code => 'IIBI'},
+         {:code_description => 'Unable to locate', :the_code => 'UTL'},
+         {:code_description => 'Located, refused exam and/or treament', :the_code => 'LR'},
+         {:code_description => 'Out of jurisdiction', :the_code => 'OOJ'},
+         {:code_description => 'Other', :the_code => 'O'}
+        ].each_with_index do |type, i|
+          ExternalCode.create!(type.merge(:sort_order => i*10, :code_name => 'contactdispositiontype'))
+        end
+      end
     end
     
     add_foreign_key(:people, :disposition_id, :external_codes)
@@ -43,7 +49,9 @@ class AddContactDispositionToPeople < ActiveRecord::Migration
 
   def self.down
     remove_foreign_key(:people, :disposition_id)
-    ExternalCode.delete_all :code_name => 'contactdispositiontype'
+    if RAILS_ENV == 'production'
+      ExternalCode.delete_all :code_name => 'contactdispositiontype'
+    end
     remove_column(:people, :disposition_id)    
   end
 end
